pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Workspace project"
requires-python = ">=3.11"
dependencies = [
    "claude-code-sdk>=0.0.20",
    "click>=8.2.1",
    "langchain>=0.2.1",
    "langchain-openai>=0.3.28",
    "networkx>=3.5",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
]

[project.scripts]
amp = "amplifier.microtasks.cli:main"

[tool.uv.workspace]
# Add all projects in the workspace
members = []

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["amplifier*"]

[tool.uv.sources]
# Example: my-project = { workspace = true }

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "debugpy>=1.8.14",
    "pyright>=1.1.405",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.10",
    "twine>=6.1.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [
    "**/__pycache__",
    ".venv/**",
    "node_modules/**",
    ".git/**",
    ".ruff_cache/**",
    ".pytest_cache/**",
    "*.egg-info",
    "build/**",
    "dist/**",
    ".data/**",
    "ai_working/**",
]
typeCheckingMode = "basic"
reportMissingImports = false
pythonVersion = "3.11"
